config: merge reader and writer
