+++ src/trx/config/priv.h
@@ -5,4 +5,3 @@
 void Config_LoadFromJSON(JSON_OBJECT *root_obj);
 void Config_DumpToJSON(JSON_OBJECT *root_obj);
 void Config_Sanitize(void);
-void Config_SanitizeCommon(void);
